--- a/editoast/src/models/train_schedule.rs
+++ b/editoast/src/models/train_schedule.rs
@@ -15,6 +15,7 @@ use crate::models::prelude::*;
 #[derive(Debug, Default, Clone, Model)]
 #[model(table = editoast_models::tables::train_schedule)]
 #[model(gen(ops = crud, batch_ops = crd, list))]
+#[model(row(derive(diesel::QueryableByName)))]
 pub struct TrainSchedule {
     pub id: i64,
     pub train_name: String,