close #10452

What si done (review commits individual):

1. Spawn tasks to dispatch cpu charge
2. Reduce train batch size
3. Avoid useless psql queries
4. Optimize a postgresql queries. Limit the number of elements in the response. Moving a filter from rust to postgres. (check perf here: https://explain.dalibo.com/plan/e7d4427b7a8ea7d5)

### Local benchmark

|               | Before | After |
|---------------|--------|-------|
| Without cache |  106s  |  108s |
| With cache    |  9.64s | 3.15s |

### Dev benchmark

|               | Before (Compression) | After (No-Compression) | After (Compression) |
|---------------|----------------------|------------------------|---------------------|
| Without cache | 90s                  |          111s          |         109s        |
| With cache    | 17s                  |          12.7s         |         8.3s        |